pyproject.toml
[project]
name = "webos-subscription-service"
version = "0.1.0"
description = "Practice dashboard for subscriber and device monitoring"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<1.0.0",
    "jinja2>=3.1.6,<4.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1,<1.0.0",
    "pytest>=8.3.5,<9.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
    "ruff>=0.11.4,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.uv]
package = false
